Phone Search Interface (#19005)
* WIP to display UI

* Working to add and remove users via phone numbers

* Only search when the number is valid

* Design tweaks

* Passing pre-commit hooks

* Storyshots

* Remove fullname parameter and pass it to prettyname per feedback from zapu

* Working to start a conversation with an unresolved phone number

* Delete alert and replace any type

* Move all css to the styles object and migrate to Kb.NameWithIcon

* Updated storyshots

* Better handling of unresolved numbers

* Refactor to remove onChangeValidity from PhoneInput

Now onChangeNumber passes two arguments: the phone number and whether or not it is valid. This
makes it possible to have code that uses PhoneInput only dispatch an action on a phone number
if the phone number is valid (otherwise there would be a race condition with setState)

* Call toJS() in mergeProps rather than stateToProps

This ensures that toJS() is only called when the props change

* Use React.useCallback to avoid re-renders

* Move results outside of the phone-input component and refactor css

* Last tweaks from Danny

* Add number pad icon [hack warning]

Note that this is a bit hacky. svg2ttf has some issue with this svg where the svgo optimized
image would be rendered by svg2ttf as a black box. By inspecting the uncompressed svg file I
saw that there was a rect over the entire image but without a fill specified. This should not
have been rendered as a black box, but svg2ttf renders it as a blackbox. It turns out svg2ttf
seems to implement its own svg rendering that is subtly broken in this case. Worked around by
just removing the rect and rendering the uncompressed svg file.

* Update icon and fix dependencies for useCallback