keybase_service_base: assure the current device can migrate
Before taking the migration lock on the server, assure ourselves that
the current device has read access to the TLF and thus can perform the
migration.  Right now it's possible for a new device, that hasn't been
keyed for the TLF yet, to get the migration request, take the lock,
and then fail to perform the migration.  This locks out any future
rekeys that would make the TLF readable, since there's nothing that
automatically releases the lock when the migration rekey fails to
happen.

Issue: HOTPOT-552