Don't emit initial intents. (#19202)
* Don't emit initial intents.

This is the wrong behavior, since this means that this will emit
everytime the app comes back from a paused state (you are switching apps
or you are coming back from background).

If anything this should only check that it's the first time ever. But
more likely it would be better to just expose these intent so JS can get
it when it's ready.

This commit just fixes the problem by removing all initial emitters (which is
fine since sharing data wasn't working anyways). A real fix would be
following the paragraph above.

* Revert "Don't emit initial intents."

This reverts commit ea249da1552bf3ba634bb6c6b6209d0b637fde03.

* Remove extra data from intent after we use it.

This makes sure we aren't emitting stale intent data.