ServiceSummary in contact search (#19173)
Add new service to fetch service summaries using user/service_maps API. These
are entirely server trust and not verified locally. ServiceSummaryMap has an LRU
cache for in-memory caching and on-disk cache using LevelDB (that's enabled for
dbcleaner to cleanup).

Service summaries are fetched during contact sync and saved with
ProcessedContact to phone book. So nothing extra has to happen during search, we
are using service summaries that we found during syncing. We are hoping to be
syncing frequently enough so these never go too stale.