chat_rpc: when loading chat convos, don't create repeated handles
When we get lists of conversations back when loading recent fileedit
conversations (or recent self-written conversations), we were always
constructing a TLF handle for each conversation, even if it was one
we've already checked (since those lists can contain the same
conversation multiple times).

Instead, do a quick check to see if we've already processed that
conversation, so we can avoid the expense of resolving the TLF name
and looking up the team settings to find the TLF ID if possible.

Issue: HOTPOT-746