kbfs_ops: finish journal initialization before edit history
This fixes a race where edit history initialization could construct an
FBO before journal initialization completed.  This could lead to a
situation where the FBO head is on the server's view, but then later
the journal is initialization to what should be the local view, and
FBO never catches up.  Then if someone tries writing to the folder via
the FBO, they'll likely either fail or cause some kind of
self-conflict.

Instead, require the code that constructs the KBFSOps to provide a
channel that will be closed when all initialization (including
journal) is completed, and only then do we let other types of
initialization commence.

I tried to think of a good test that would fail before this commit but
succeed after (without having to be significantly rewritten) and I
failed.

Found while debugging a (maybe?) unrelated issue, HOTPOT-803.