init: allow a "kbfs.mode" config option
A few users on slow desktops wanted the ability to run KBFS in
"constrained" mode.  They currently have to override the systemd
definition for the kbfs service if they want to do that on Linux, and
I don't think there's any way to currently do it on other platforms.
So just put it under the "kbfs.mode" path in the config file, to make
it much easier to do.

Issue: HOTPOT-839