tlf_journal_test: vdebugging + ignore post-shutdown state changes
I wasn't able to reproduce a CI test flake, and there wasn't enough
info in the logs to figure out exactly what happened.  But in the
flake, the test had finished successfully and was trying to clean up,
though the journal was trying to react to some kind of unknown work
signal that was probably some rare flake.  So:

1. Turn on vdebugging for TLF journal tests.
2. Ignore attempted state changes after the test shuts down.

Issue: HOTPOT-725