journal_manager_test: wait for pause to take effect
The following race caused some test flakes:

1. Test creates a journal, spawning background goroutine.
2. Test pauses journal by sending a pause signal.
3. Test adds block to journal and sends a work signal.
4. Background goroutine selects the work signal before the pause
   signal, and flushes the block.
5. Test restarts the journal manager.
6. The journal is empty, and gets immediately nuked, and doesn't show
   up in the count, failing the test.

The problem is that the test doesn't wait for the pause to take effect
before putting the blocks into the journal.  So this commit adds a way
to create delegates in the journal manager tests, and uses that to
make sure the pause has taken effect.

Issue: HOTPOT-819