Fix dropped notifications on android
On Android, `HandleBackgroundNotification` is the *only* thing that will
lead to a notification. It's possible that both the silent and nonsilent
notifications are stale–Say if you haven't had an internet connection
or Android/Google delays the notification.

The reason the code was like this is because on iOS the notification
works differently. It's possible the silent notif was delayed, and the
OS already showed the nonsilent version. In that case we doe a staleness
check to make sure we don't show duplicate notifications.

On Android, duplicate notifications isn't a problem. We check in the
Java code wheter we have already seen a notification for that
conversation and message. We can safely ignore the stale logic on
Android.