tlf_journal: use ctx timeout when putting blocks, if available
The journal has a built-in timeout of 11s when trying to put a block
into the journal.  (This includes the waiting time if the journal is
full and needs to drain before putting the block.)  But if we're
writing via SimpleFS (which puts a timeout of 1 minute on every
operation), there's no reason to artificially constrain the timeout.

So instead, only use the default timeout if none is currently set.  If
one is set, use 60% of it (default macOS timeout of 19s * .60 =~ 11s,
which is the current timeout) if that's bigger than the default
timeout.

Issue: HOTPOT-904