cr_chains: set final path on rename op even when removed
A few users have hit this sequence in a single MD update:

* Rename a file 'foo' to 'bar'.
* Delete 'bar' (either through an `rm`, or due to renaming something
  else over `bar`).

Then when trying to construct the edit notifications for this revision
(either when flushing it to the server, or when reading it from the
journal to construct the "unflushed" notifications), the code panics
because the rename operation is missing a final path.  Extra debugging
added for #19405 revealed this pattern in a new log.

The op's final path is missing because, when reverting rm+create op
pairs that come out of the chaining process into rename ops, we were
only looking for the final path of the corresponding create op.
However, in this case the create op is thrown away since the whole
file/dir is removed in the end, and we went ahead and included the
rename op anyway, without a final path.

The solution is to try to find the path from the rm op instead, which
should still be around.

This also adds a regression test to catch this case, as well as
`folderBranchOps` code that strips out the final paths before the
notification process while in test mode, to make sure the tests match
the app cold-start behavior.

Issue: HOTPOT-803
Issue: #19405