folder_block_ops: avoid panic when undoing node cache ops on error
When node cache operations return errors (e.g., due to other bugs such
as HOTPOT-814) and the pointer state is inconsistent, we might need to
execute undo functions.  But those undo functions themselves can
legally be `nil`, and in those error cases are even more likely to be
`nil`.  So Check that before executing them (which is consistent with
the way undo functions are used in other parts of this code).

Issue: HOTPOT-948