init: don't trigger cache warnings in single op mode
If a user sets their KBFS into constrained mode, it won't run a
service for the remote cache proxy.  Then whenever they use git, they
would get a warning error about not being able to initialize the disk
cache in the git process, which is trying to connect to the remote
proxy.

Instead, just suppress the warnings, I think it's fine for the git to
not be able to use the remote cache in that case.