kbfs_ops: helper func for synced TLFs that filters out bad TLFs
If the current user can't read a particular synced TLF, that's ok
because they might not have access to it -- e.g., it could have been
synced while logged in as a different user on this same device.
That's ok, so just skip TLFs like that when listing favorites and the
full sync config list.  Do so using a helper function to look up the
TLF node MDs and handles, while skipping inaccessible TLFs (after
logging the error).

Issue: Y2K-329