keybase_service_base: cache negative writership for a short time
We use cached team info when possible, unless we're checking for
membership of a user that's not present in the cached info.  In that
case, we force a reload, just in case there's a race with the
membership-changed notification from the service.  Usually this is
fine because it's rare to do a negative lookup multiple times in a row
-- after the first error, the client will stop.  However, in the case
of stat-ing all the files in a public directory, we call this function
over and over again to figure out if the current user is a writer (to
set the writer bit in each entry's permissions), and since it's a
public folder the user is not (usually) a writer, so it was forcing
the service to reload the team every time, slowing things down
considerably.

Instead, cache the fact that a user is not a writer for a short time
(5s).  This is safe with respect to the FUSE calls that return
permissions on a file (which tells the kernel to cache that info for
one minute anyway).  However it potentially sets up a race with any
situations where multiple notifications race with each other when the
team membership changes and we need to figure out whether the current
user is now a writer.  I can't think of any such situations right now,
but we should keep an eye on it.

Issue: HOTPOT-1009