key_manager_test: sync device 3 after rekey during test
A CI failure (that I couldn't repro locally) showed that a device that
should have been able to read a TLF after a rekey, wasn't able to.
From the log I think it was still trying to read an old version, so
I'm fixing it by syncing before trying to read it from that device.

Issue: HOTPOT-1059