test: re-sync the active TLF after adding a new assertion
libfuse handle TLF name change invalidations in a background
goroutine.  Before this change, if that goroutine was too delayed,
then after a new assertion is added, the TLF name might not have been
invalidated by the time the test expects needs to be (e.g., the next
`in...TlfNonCanonical()` call).  With this change, we just sync the
active TLF after a new assertion is added, which waits on the
invalidations caused by the new assertion to finish.

(This is based on a CI test failure that I wasn't able to repro
locally, but it's my best given from the logs.)

Issue: HOTPOT-1100