journal: return error when local conflict view doesn't exist
If someone tried to access a local conflict view after it was
finished, KBFS would panic because it was using the cleared
(i.e. null) TLF ID from the journal manager instead of a real ID.

Instead, make it clear that the view is gone in that case by returning
a proper error, and add a test that reproduced the panic before the
fix.