kbfsgit: send LFS progress messages on uploads/downloads
LFS allows us to send progress messages during uploads or downloads,
by sending a JSON message containing the OID of the operation, plus
the number of bytes processed so far.

For downloads, this is pretty easy, as we can just re-use the strategy
of wrapping the io.Writer to send progress messages, just like we do
for normal git.

It's a bit trickier with uploads, because uploads have two stages --
one copying it into the journal, and one flushing it from the journal.
LFS only lets us have one progress-reporting stage, so we choose the
journal flush since it's usually the slower of the two.  Unfortunately
this means a bit of a lag in reporting since we wait until the copy
finishes before reporting the journal status, but that's good enough
for now I think.

The size of the file in the journal doesn't match the plaintext size,
so we need to scale the upload progress messages to reflect the
plaintext size reported by the LFS CLI.

Issue: HOTPOT-1275