folder_branch_ops: disallow empty file names or links
A user hit an issue where they created a symlink with a blank name
using SimpleFS, which then caused all sorts of rendering errors in the
GUI and `ls` problems on the command line.

Just disallow the creation of names like that, and add a SimpleFS test
for symlinks.

Also add an error stacktrace to the disallowed prefixes/names error,
for easier debugging.

Issue: HOTPOT-1276