folder_block_ops: don't prefetch for "permanent" cached blocks
When a user reads a block that has a "permanent" lifetime (which means
it can't be kicked out of the transient cache, usually because it is
currently being written by the user and has unflushed data in it), we
shouldn't kick off prefetching.  This just ends up resulting in a
(non-fatal) error in the prefetcher because it doesn't expect the
lifetime to be permanent along that path.