tlf_journal: protect against crasher during `resolveBranch`
A user hit the following issue:

1. Their client completed a large conflict resolution, and called
`tlfJournal.resolveBranch()` to finalize it.
2. The new commit was written to a new mdJournal directory, which was
then swapped into place, making it the main MD journal.
3. Then, KBFS restarted, _before_ `resolveBranch()` had a chance to
mark all the MD revision markers in the block journal as ignorable.
4. After the restart, the journal resumed flushing blocks.
Concurrently, the user wrote more revisions into the MD journal,
re-using some of the same revision numbers that had previously been
conflict-resolved away in step 1.
5. While flushing blocks, when the TLF journal discovered a revision
marker for an old MD revision (which had originally been resolved and
squashed in step 1), it decided it was ok to flush a _new_ revision
with that same revision number.  This resulted in the revision being
live to other clients _before_ the blocks associated with that
revision were available, and other clients could not read the TLF.

To protect against such a race, this commit adds a new notion of a
"journal ID" for the MD journal.  Each time the MD journal is cleared
or resolved, and then a new revision is written into it, it gets a new
unique ID that is persisted to a file.  Also, each MD revision marker
gets labeled with the current MD journal ID.  When flushing, the code
ignores any MD revision marker that doesn't match the current MD
journal ID.  That way if KBFS crashes/restarts after clearing the MD
journal, but before updating the block journal, the old MD revision
markers will be properly ignored when the flush resumes.

Issue: HOTPOT-1553