git: don't use singleOp in the journal for LFS uploads
If a user tries to upload giant multigig files using LFS, we don't
want them to be uploaded in a single MD update to the server, because
it increases the risk that something will go wrong or take too much
memory in conflict resolution.  Since LFS doesn't have the same
locking requirements that regular git has, we can easily upload those
files using the normal journal flushing procedures, without having to
squash everything together first.

To do that, this commit changes the `JournalManager` to remember a
default background work mode (that's already provided in the KBFS init
params), and to use that whenever enabling journals.  Then,
git-remote-keybase just needs to set that correctly when starting up
in LFS mode.

Issue: HOTPOT-1554