Fix PGP kid reuse breaking sigchain (#21733)
PGP keys are not uniquely identified by KID. KID is only generated by
hashing the primary key public key material. The rest is malleable.
Different uids, subkeys, and signatures will yields the same key but
also the same PGP fingerprint. However, it is possible to generate new
PGP key with the same key material but different creation time. This
will yield a new PGP fingerprint but same Keybase KID.

There was an issue where we weren't properly setting `ActivePGPHash`
when new PGP key was being delegated after key with the same KID was
revoked.

During sigchain parsing, we would go through:
1) Delegate key A with kid=X - new key, ActivePGPHash is set to A.
2) Revoke kid=X. keys[X] is set to Revoked, and ActivePGPHash to "".
3) Delegate key B with kid=X - there already is X, set status to
	Uncancelled (forgot about ActivePGPHash here).

This resulted in sigchain not being replayable because we were taking
PGP key A to verify reverse sig of PGP key B in link 3.

The fix is to set `ActivePGPHash` in ComputedKeyInfos.Delegate when CKI is
already found for that KID, when we are already "uncancel" that CKI, and refresh
its CTime and ETime.