pullTranscript: do bunch of pulls with pagination
We only want messages for current user and reported user. Add filtering
and skip messages coming from other users.

In busy teams, we likely need to do a bunch of pulls until we get
requested amount of messages for transcript (set via const to 50).

So we are no longer just pulling 100 last messages, instead we want at
most 50 messages from selected users, and we do 5 pulls 100 messages
each to find them (or not).