autogit: avoid possible panic
A user hit a panic below this code a long time ago, because of a `nil`
`b.repo`.  I couldn't repro it or see how it was exactly possible, and
the code has changed a bit since that version so it might not be
possible anymore.  In any case, it seems like a safe check to have, so
might as well add it in.

Issue: HOTPOT-1697