tlfhandle: when resolving by ID, don't ignore network errors
A user tried to resolve an implicit team by ID when their network was
flaky.  They got back an `AppStatusError` with an `APINetError` code.
This caused them to fallthrough to the code that treated it like a
normal team, which unfortunately set the name of the TLF to the
`__keybase_implicit_team__`-style name, which caused a lot of
confusion until the next reboot.

We should treat network errors like that as real errors, and only
fallthrough for certain kinds of errors.  There's not a proper error
type yet for that, so this commit does our best to improve the
situation for now.

Issue: HOTPOT-1698