indexer: launch a full index of a synced TLF
When a full sync of a TLF starts at a specific revision, it sends a
message to the indexer.  The indexer waits for the sync to finish and
then does a full index.

This still probably needs a few tweaks to make it production ready,
and those will come in future commits.  For now the indexer isn't
enabled in production, so it doesn't matter much.

Issue: HOTPOT-1494