Crypto in GUI fixes (#22247)
* Allow updating output if no output is present.

Entering text, setting a recipient, and then clearing the recipient did
not re-run the operation because outputMatchesInput = true.

In addition to checking if the store input matches the RPC input, we
also check that the store has output. If no output exists, then update
the store's output with the result of the RPC (even if the inputs
match).

* Rename noIncludeSelf to hideIncludeSelf

* Handle force signing for SBS recipients

* Fallback to 'cannot verify' message for 1002 SC

* Easier fix for clearing clients

(cherry picked from commit ef2714a934c873ab1d96507f6e0fe55d80810de5)