indexer: track whether a directory is fully indexed or not
We no longer need to track whether a given doc ID is used or not (this
is pre-determined by whether we are newly-indexing something, or
updating something).  But we do need to track whether all the children
in a directory are fully indexed, since if it is, we can skip
re-indexing that whole directory after resuming an interrupted
indexing.

Issue: HOTPOT-1763