md_util: helper method for getting changes for an MD range
The indexer needs a list of simplified changes that happened to files
(just writes, renames, and deletes) after one revision, up to another
revision.  The needs here are a bit different from the edit history,
because it's best that it be a collapsed list of changes in the
revision range (to avoid indexing the same files multiple times, or
indexing something when we know it will be deleted), and also needs to
include BlockPointers.

So this adds a helper funcion that uses `crChains` to collapse the ops
for the revisions, and creates simplified `ChangeItem`s for those ops
that the indexer will be able to consume.

Issue: HOTPOT-1495