folder_branch_ops: update pointers for new files correctly
Previously, we were updating the pointers for newly-created files
_after_ all the other pointers were updated, and the deferred writes
were replayed.  This is wrong, because those deferred writes need to
operate on the new pointer, and if there were deferred writes for the
new file, they would get lost due to trying to write to the old
pointer.

Instead, make sure the pointers are updated first, during the same
phase as the other pointer updates, and make sure deferred writes
refer to the new pointer.