folder_block_ops: block writes while copying dirty dir blocks
The following scenario was causing bugs when syncs and writes were
happening concurrently:

* `SyncAll` starts.
* It gets the list of dirty dirs, which sets
  `folderBlockOps.dirtyDirsSyncing = true`.
* Then a write to a file sneaks in.  Because `SyncAll` hasn’t yet
  started syncing the file, the update to the directory is deferred,
  but not the write to the file itself.
* Then `SyncAll` starts syncing the file.
* By the end of `SyncAll`, the file has synced and so its dirty status
  is cleared.
* Then the deferred directory update is replayed, causing the
  directory to be dirty without any corresponding dirty file or
  directory operation.
* After that, the next `SyncAll` starts, but it doesn’t consist of any
  operations.  It’s is only one dirty directory, without any
  operations to support it.
* This confuses the `folderUpdatePrepper` code, which decides that the
  directory hasn’t actually been updated because it’s not connected to
  an operation, and it turns the directory pointer's update into just
  a single reference in the MD's list of operations.
* Because of that, we don't update the node cache's pointer, and the
  directory's node remains pinned at its old block ID.
* Then the next write to a file in that directory then results in an
  error because the expected block pointers don't match up between the
  directory's entry for the file and the file node itself.

This happens because we allow writes to happen during syncs, so we
don't slow things down for the user very much.  But this makes things
complicated, since directory entries can change during the sync.  So
instead, this commit blocks writes just during the sensitive phase of
sync that is copying the dirty directory blocks.  That way, in
constrast to the above, the written file won't be included in the
sync, and its write will be deferred along with the directory update.
Then the next sync will correctly pick up both the file write and the
directory update.

The cost of this is that writes will be slightly buffered during a
sync, but during that block there shouldn't be any disk or network
accesses, so it shouldn't affect performance too much overall.

Note that I couldn't figure out a good way to write a reliable unit
test for this scenario; I think it would require to much test-specific
scaffolding to be inserted into the production code, and it doesn't
seem worth it to me.  I was able to reliably reproduce it using the
new KBFS search code though, and I'm reasonable confident it's fixed
now.

Issue: HOTPOT-1903