libfs: support `.kbfs_dirblock_` virtual directories
This lets a user directly access a directory given its block pointer.
This is useful in bug scenarios where the parent or root directory of
a desired directory is corrupt or unavailable, but we still want to
read a subdirectory.  We can get the block pointer from the
`.kbfs_update_history` file or from the logs.  For example:

```
ls .kbfs_dirblock_02002e7771ea717335454c43dd1a8999d1d6dab20c73fb4372b7df748800195dfe.20.1.ef2e49961eddaa77094b45ed635cfc00.direct
```

This presents a read-only version of the directory block.

If you happen, for some reason, to use this for a directory that's
already accessible, it will turn that directory read-only because they
will share block pointers in the node cache.

Issue: HOTPOT-1986