kbfsgit: don't include symbolic refs in for-push list
When a mirror push (`git push --mirror`) asks a remote helper for its
list of references, it apparently doesn't expect to get back symbolic
references.  If it _does_ get back a symbolic reference, it will check
the local clone where the push is being done, and if there isn't a
symbolic ref, it will delete the one in the remote repo (the one in
KBFS, in this case).  This seems to be because symbolic references are
not transfered between remotes; it's not part of the network protocol.

For an example of an "official" remote helper demonstrating this, see
the "remote-curl" helper in the official git repo:
https://github.com/git/git/blob/51ebf55b9309824346a6589c9f3b130c6f371b8f/remote-curl.c#L232

Issue: HOTPOT-1970
Issue: #22520