Fix TestAccessRequestIgnore test (#22813)
Use SetTestingNoCachingMode(true) in tests where a change to a user is made and
that information is then queried using uidmapper really quickly.

The cause for this and related uidmapper bugs / flakes is that API servers have
uidmapper caches as well. That cache is cleared with user.identity_change
pubsub message, so when someone does something on one API server, e.g. calls
profile-edit, all of the API servers should have their uidmapper cache for that
uid cleared afterwards.