cr_chains: avoid `revertRenames` panic
A user hit a panic here, where `newChain` was `nil`.  Their system
seemed to be in a generally bad state -- some of the Nodes seemed to
be behaving inconsistently.  But the log wasn't big enough to fully
debug why; it could be related to other issues we've already fixed.

Just in case this happens again, avoid the panic here by returning an
error with more info.

I'm not 100% sure this is a good idea.  It might be better to have the
panic to surface the problem sooner rather than paper over it.  For
now I'm leaning towards getting rid of it though, to keep the client
more stable, since it probably only affects the edit notification
history.

Issue: HOTPOT-2029