config_local: don't clean unsynced TLFs when sync DB wasn't loaded
Sometimes a temporary storage error prevents us from reading the
synced TLF DB on startup (especially on iOS).  Make sure we don't try
cleaning the unsynced TLFs in that case, because we might start
throwing desired blocks away.

Instead, try reloading it for a while in this background goroutine.
If it really can't be loaded, give up.

Issue: HOTPOT-2092