transactions.go refactoring (#22758)
Introduce AddMemberCandidate struct in transactions.go. It functions as the
intermediate data object when adding people to a team. Its instance is returned
by calling ResolveUPKV2FromAssertion with an assertion. It can be then
inspected - it contains information about parsed input assertion and whether it
resolved to a user or not. After that, AddOrInviteMemberCandidate can be used
to create signatures needed to add person represented by the instance of
AddMemberCandidate to the team.

Additional changes include:

Change UserRolePair.AssertionOrEmail to Assertion. Raw e-mails can't be passed
through that anymore for some time now.

Since around the time when bot support was added, some operations are no longer
atomic, that is, they can fail and leave transaction in broken state. Handle
this better by new error type and saving the error in transaction struct so it
can't be Post-ed later if it happened.