tlfhandle: handle identify errors correctly on single assertions
When identifying a single assertion before resolving it (e.g., because
it contained a "+" character), and when extended identify behavior was
set, we weren't calling `makeTlfBreaksIfNeeded()`.  So if there was an
error, reporting the error would hang because there was nothing
waiting on the extended identify `userBreaks` channel.

So this commit wraps that behavior into a util function and makes sure
it won't hang.

Issue: HOTPOT-2141