disk_block_cache: force leveldb compaction after big deletes
It turns out that goleveldb runs its table compaction after a certain
number of seeks happen, but it doesn't run it after a bunch of deletes
happen (as far as I can tell).  That means when you for example
disable a synced TLF, it will mark all the blocks as deleted, but it
will continue to use up the disk space until some other compaction
event happens.  And if you've deleted ALL your blocks, that compaction
event will never happen, because there won't be any more seeks.

So, this forces compaction to happen on large-ish delete events.

Issue: HOTPOT-2204