search: improve indexing memory usage
Bleve's indexer takes a lot of memory proportional to batch size.  I
asked about this on blevesearch/bleve#1347, but haven't yet found a
way to reduce memory usage besides just making the batch sizes
smaller.

Unfortunately, making the batch sizes smaller increases CPU usage (and
indexing time) so we don't want to do this unless we really have to.
So this commit is a first attempt to reduce the batch size in the case
that there's not much free memory available.  We change the size for
each batch, depending on the amount of available memory.

Also, reduce the size of the clean block cache for the indexer.  Don't
reduce it too much though, because it would cause a lot of disk and
decryption load if it's too small.

Issue; HOTPOT-2205