Clearing operation store on unmount
Since the input and output views are split on native the values were not
cleared when swiping forward/back. This commit adds a useEffect hook to
dispatch CryptoGen.resetOperation when leaving the input screen. This
works quite well so far.

Additionally changes to OutputInfoBanner were made to accommodate null
children.