Use sound and LED for notifications on Android. (#22963)
According to some previous bug reports, the client used to have sound
and lights for notifications on Android, but stopped doing so at some
point. This commit restores sound and lights for notifications -- the
lights uses the system defaults for notifications, and the sound will
either be the system default or the resource file specified in the
event depending on how the setting is toggled.