libkbfs: put fetched archived blocks in the sync cache
If a block has been mistakenly marked as archived on the server due to
an old KBFS bug, but it is still reachable through the master view of
the TLF, and that TLF has been configured to sync on the local device,
we still want to put that archived block into the disk sync cache.
Otherwise, the prefetcher will get confused and prefetching will never
complete.  (Also, the data wouldn't be available offline.)

This would be a relatively simple fix to `bserver_remote.go`, except
that we _don't_ want to cache these archived blocks if we're reading
an old revision of the TLF where the blocks are supposed to be
archived.  In that case, we don't want to set the cache type for the
fetch to be the sync cache.  So all the places that set the cache type
now need to be aware of the "branch name" of the local view of the
TLF, instead of just knowing the TLF ID.

To accomplish that, we pass the branch name into the `BlockOps.Get()`
call, and add a new block request action type to indicate that the
fetch is for a non-master branch.

I punted on passing the branch name around for the specific case of
fetching MD blocks for a historical version of a TLF -- the call stack
size was too overwhelming for such little benefit, for now.

Issue: HOTPOT-2253