modes: delay background synced TLF initialization for search mode
If a synced TLF is initialized too early, the indexer might not yet
have registered to receive notifications about synced TLFs.  So just
delay that background work for a few seconds.  A better, future fix
might be to have the indexer itself check for any missed notifications
after it has registered itself, but that's a lot more complicated.