disk_block_cache: evict blocks in bigger batches when possible
A user had an issue where their syncs were blocked when suddenly their
disk filled up (I think by some external thing) and their disk block
cache was suddenly over capacity.  A `SyncAll` started, while they
were still doing concurrent writes.  I think one of the writes did a
Put into the disk block cache, which caused a bunch of evictions while
locked.  So many evictions that the `SyncAll`, which I suspect was
blocked on the disk block cache lock (trying to read a block), timed
out.

The best solution is probably to release the lock once in a while
during eviction, but that seems a bit complicated.  So for now, I'm
just trying to calculate a good number of blocks to evict in a single
batch based on the average block size, instead of always trying to do
10 at a time.  Hopefully this speeds up the process and makes it less
likely to block for long periods of time.

Issue: HOTPOT-2281