libfs: move .kbfs_profiles support into node wrappers
So they can be accessed via SimpleFS, under both `/keybase/` and
inside TLFs.  And the libfuse support was refactored to use all the
libfs stuff.

This is a bit more complicated than other `.kbfs_` special nodes,
because it's a directory.  That means we have to build a full file
system interface for that, so that `folderBranchOps` and `rootFS` can
use it, and we have to extend the SimpleFS support for the root FS to
include special subdirectories.

libdokan still needs some work to support timed profiles; I'm leaving
that work until later if we decide it's necessary.  (But now we can
always get those timed profiles out of SimpleFS anyway.)

Issue: HOTPOT-2434