folder_branch_ops: unregister from MD updates on shutdown
If someone calls `FinishResolvingConflict` on a TLF, it will shutdown
the existing `folderBranchOps` for that TLF.  Normally that's fine if
it's really a conflict TLF, since it shouldn't be re-used again.  But
if they call it on a normal TLF, they'll receive an error and life
will go on.  But the FBO will still get shut down.  Then the next time
the TLF is accessed, a new FBO will be made and try to re-register for
mdserver updates, which will cause the panic.

So, just unregister during shutdown so this doesn't happen.

Issue: HOTPOT-2495