prefetcher: proxy wait channel requests while paused
If the prefetcher is paused because it's on a cellular network, or the
app has been backgrounded, we shouldn't block the "wait channel"
requests indefinitely.  These are processed on the same queue as
regular prefetch requests, and so can't be processed until we're
unpaused.  But the caller is just asking for a channel they can wait
on; they don't actually expect this call itself to block.

So if the prefetcher is paused, return a proxy channel that the caller
can use, and launch a goroutine that will close this channel when the
request is eventually processed.

I considered _always_ returning a proxy channel, but it seemed nice to
still wait on `ctx.Done()` and the shutdown channel in the case where
we expect the request to be processed quickly.

Issue: HOTPOT-2569