remove the gconf dependency on Arch Linux (#23909)
* update stale fields in DOT_SRCINFO.bin.in

In the usual Arch Linux workflow, this is a generated file. However,
generating it requires the `makepkg` tool, which is difficult to run on
anything other than Arch Linux. To avoid that dependency, we've checked
in this generated file as another template. However, as a predictable
consequence, it's gotten out of sync with the PKGBUILD file it's
supposedly generated from. These changes bring it back into sync.

One alternative to this whole setup, would be to just run a Docker image
for Arch Linux, and to legitimately generate this file. If that would be
convenient in the current build environment, it's worth considering.

* remove the dependency on gconf

This was suggested in
https://aur.archlinux.org/packages/keybase-bin/#comment-723637.