keybase_service_base: don't ignore deleted errors if missing name
Usually when the service returns a `UserDeletedError`, it still fills
in some details about the user.  There seems to be at least one case
where it doesn't though, and in that case want to return the proper
error back up to the GUI so that it can handle it correctly and
display the right banner.

Issue: HOTPOT-2557