favorites: don't refresh on mtime when last ID was the same
There's no point in refreshing the favorites list when the mtime of a
TLF changes, if that's the same TLF that we refreshed for previously.
We only use the mtime to relatively order TLFs among themselves, so if
the same one changes again, that won't change the relative order.

Issue: HOTPOT-2582