Team management add members: handle people already in the team (#24033)
Add FindAssertionsInTeamNoResolve RPC that checks if any assertion from a list
of assertions is already invited to the team or added as member.

It's called by add member wizard in the GUI to be able to skip some assertions
after using one of the mechanisms (team builder, bulk emails, bulk phone
numbers) to queue new members in the wizard. When any of the assertions are
already invited or in the team, there is also a message displayed in the
wizard.