AddMemberTx: AllowRoleChanges option (#24137)
Add `AllowRoleChanges` boolean field to AddMemberTx. When set to true,
functions adding crypto members to the team will no longer fail with
ExistsError when the member already exists in the team with current UV. When
the requested role is different the current role, it will queue an add with
that role.

Note that this only works for crypto members (Keybase users with PUKs). For
now, the transactions code does not support changing role of a Keybase-type
invite - which would require canceling old invite and adding new one using an
`team.invite` link. So it's technically legal within current sigchain rules,
just not supported by the transactions.go code.