kbfs: don't start archiver bg goroutines until SFS is initialized
If the archiver starts making SFS calls before the `MDOps` instance is
set in the config (which happens during an asynchronous init process),
it could try to lookup TLF handles and IDs too early, resulting in it
thinking it needs to create new TLFs over existing TLFs.

Instead just wait for that part of init to finish with a dumb loop.